+++ src/fields/AddressField.php
@@ -635,4 +635,26 @@ class AddressField extends Field implements PreviewableFieldInterface
 //        ProximitySearchHelper::modifyElementsQuery($query, $value, $this);
     }
 
+    // ========================================================================= //
+
+    /**
+     * @inheritdoc
+     */
+    public function getSearchKeywords(mixed $value, ElementInterface $element): string
+    {
+        /** @var AddressModel $value */
+        return implode(' ', [
+            $value->formatted,
+            $value->name,
+            $value->street1,
+            $value->street2,
+            $value->city,
+            $value->state,
+            $value->zip,
+            $value->neighborhood,
+            $value->county,
+            $value->country,
+        ]);
+    }
+
 }